KUDU-2612 p2 (b): add transaction status retrieval
After offline discussions with Andrew, it became clear that TxnManager
should provide an asynchronous interface to commit a transaction, i.e.
something similar to CreateTable()/IsCreateTableDone().  To implement
that, the TxnManager needs to check for the status of the transaction
after initiating the commit phase by issuing corresponding call
to TxnStatusManager (that's implemented as CoordinateTransaction() RPC
to TabletServerAdminService with BEGIN_COMMIT_TXN operation type).

This patch introduces the required server-side piece to retrieve the
information on a transaction status from the TxnStatusManager.  I'm
planning to introduce corresponding bindings via the TxnSystemClient
in a separate changelist.

This is a follow-up to efd8c4f165460b7fa337b8ebd1856b10bc274311.

Change-Id: I45f099d943f2b7955d6d561a1cb883343c7b79a4
Reviewed-on: http://gerrit.cloudera.org:8080/16495
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins