[kuduraft] region_durable_index maintainance in consensus_queue
Summary:
A new region_durable_index is defined and maintained by consensus_queue.
An index is considered to be region-durable if the OpId is replicates atleast to
one additional region other than the leader's region. This index is propagated
on every request (through ConsensusRequestPB message) sent by the leader (or
proxy) and the follower syncs his own view of region_durable_index with the
leader's view of this index. This is useful for upper layers to implement purge
(or garbage collection) policy for the log files. Note that this index is useful
only when the raft ring is configured to have nodes in multiple regions (as
defined by RaftPeerAttrsPB). The index is exposed to upper layers through
RetentionIndex (RaftConsensus::GetRetentionIndex)

Test Plan: tests to be added in mtr and fbcode side

Reviewers: arahut

Subscribers:

Tasks:

Tags: