KUDU-3201: [hms] Support gzipped HMS notifications (GzipJSONMessageEncoder)
This patch adds support for HMS notification messages encoded
via the GzipJSONMessageEncoder that was added in
HIVE-20679. It does this by decoding the base64 encoding
and decompressing the messages sent with the `gzip(json-2.0)`
format.

Because HIVE-20679 is only in Hive 4 which is unreleased,
I created a static encoded message to test the gzip functionality.
I also  manually tested this functionality on a real cluster that has
HIVE-20679.

While testing this patch I found and fixed a bug in our zlib.cc
Uncompress method. Without the fix zlib would
return a BUF_ERROR when the decompressed data fits
in the output buffer on the first call to `inflate`.

For reference I used the following post to help understand
the expected usage of `inflate`:
https://zlib.net/zlib_how.html

Change-Id: I0f94e2446255e1fafb1dac9bf9ce23b81d6c0960
Reviewed-on: http://gerrit.cloudera.org:8080/16565
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>