[master] introduce filter_type for ListTables
This patch introduces a new 'filter_type' field for the
ListTablesRequestPB message.  With this patch, it's possible
to explicitly filter tables to be returned from the system catalog.

I added corresponding unit test as well.

This patch doesn't update kudu CLI -- this is to be done
in a separate patch.

Change-Id: I9d81ad740d2fc1bad0d1ce8a26cb9b9c6429c786
Reviewed-on: http://gerrit.cloudera.org:8080/16561
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins