[hms] Fix testSyncDisabled on Hive 4
In Hive 4 the MetastoreDefaultTransformer ensures that
external tables don’t use the warehouse as a their storage location.
Kudu doesn’t actually use this location, but the location must still be
valid.

This patch changes the test to re-fetch the table after creation
to get the location after translation has occured. It also deep
copies the table before altering to ensure any unwanted properties
are not copied over.

Change-Id: I2dc12f350794314325736be581888cb69be917a1
Reviewed-on: http://gerrit.cloudera.org:8080/16574
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>