[client] robust handling of absent master address(es)
This patch updates Kudu C++ client with robust handling of the absence
of master addresses prior to connection attempt to a Kudu cluster.

Before to this patch, KuduClientBuilder::Build() would hang in
ConnectToCluster().  After this patch, KuduClientBuilder::Build()
instantly returns Status::InvalidArgument() if no master addresses
are provided.

This patch also contains a test for the new behavior.

Change-Id: I4c7400d09298c01cd0b349cc2b9f481d1214f148
Reviewed-on: http://gerrit.cloudera.org:8080/16589
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>