[kudu-admin-test] fix TestUnsafeChangeConfigWithFiveReplicaConfig
I saw the AdminCliTest.TestUnsafeChangeConfigWithFiveReplicaConfig
scenario failing pre-commit TSAN build [1] with the error like below:

  Timed out: Committed consensus opid_index does not equal 1
      after waiting for 30.014s. Last opid: 2.2

As I understand, the reason for failing the assertion is an extra
election round that has happened after startup of a tablet.  This patch
addresses the issue in the TestUnsafeChangeConfigWithFiveReplicaConfig
and other scenarios which might be affected if same ever happens again.

[1] http://jenkins.kudu.apache.org/job/kudu-gerrit/22585/BUILD_TYPE=TSAN

Change-Id: I1e77c440b872a42c68b6d7d1cb0563fee0470fc2
Reviewed-on: http://gerrit.cloudera.org:8080/16601
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>