[java] Fixes for JDK 12 to 15
This patch fixes various issues found when testing
with JDK 12 to JDK 15.

The main fix is changing `HostAndPort.toString()`.
`InetSocketAddress.toString()` changed in JDK15
to include “unresolved” in the ouput.

I also needed to fix a few SpotBugs issues and
disabled `kudu-jepsen` on JDK 12+ given
Clojure only supports JDK 8 and 11.

The Hive tests do not run due to HIVE-21508. An
upgrade to Hive 3.2.0 once available should
resolve that issue.

Change-Id: Ie992d089a06e8d4afecda39a234e19a45730381e
Reviewed-on: http://gerrit.cloudera.org:8080/16600
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>