[kuduraft] memory management for index chunks
Summary:
kuduraft maintains dense index file for random read of replication log. This
dense index files are implemented as index chunks. Each entry in the index chunk
contains the offset of a particular OpId in the replication log. For faster
access (reads and writes), these index chunks are mmapped. By default, each index
chunk is configured to have 1000000 entries (i.e 1 million OpId entries) and the
size of an index chunk is 24MB. These index chunk files are deleted when the
replication log files are purged. Upstream kudu has GC thread which purges
replication logs and also the index chunk files periodically, but it is disabled
in kuduraft fork (to give more control of purging of replication logs to the
upper layer).

However, this could lead to scenarios where hundreds of index chunks could get
mmaped leading to substantial increase in rss memory consumption by the process
until the replication logs are purged. This diff tries to address the problem
by having the ability to dynamically mmap and unmap the required index chunks
and not tying the mmap/unmap of the chunks to the open/deletes of the
underlying chunk file.

The max number of entries that could be mmapped at any given time is
configurable through 'kNumChunksToMmap'. The latest chunk is always pinned
(i.e always mmapped) - the rationale here is that writes always happen to the
latest chunk. If a lagging peer asks for an OpId in an index chunk that is
currently not mmapped, then the index chunk is mmapped dynamically. A 'victim'
chunk might have to be unmappedbefore this-and the victim is always the oldest
chunk that is still mmapped.

Some observations:
1. Chunks are only 'read' when a peer requests for an OpId that has fallen off
of the leader's log-cache (and most likely the peer is lagging). Given this,
'kNumChunksToMmap' can be configured to a very low value (1 or 2) on peers which
will never serve read requests (primarily every peer other than the leader). In
future, we could even enhance this so that kuduraft automatically adjusts this
on every leader election/config-change
2. If there are multiple peers that are lagging across OpIds that are a million
OpIds away from each other, then it could lead to thrashing (frequent and rapid
mmap and unmap of index chunk files). For example, follower1 is trying to catch
up OpId (1, 1), follower2 is catching up (1, 1000001), follower3 is catching up
OpId (1, 2000001) and so on. However, this seems like a rear scenario. One
solution could be to configure each peer to have its own 'slot' in the leader's
mmapped index chunk. This could be enahnced in future if needed. Obviously all
of this will not be needed if the ultimate goal is to move away from dense index
files to sparse index files.
3. upstream kudu seems to have moved away from mmapping index chunk files and
rely on pread/pwrite to do random access of the index chunk. This is done
through interfaces in env_. We could try to incorporate it whenever we rebase
onto latest upstream kudu
4. Even though the chunk files are dynamically 'unmapped', the underlying file
itself is not closed and continues to be tracked through 'open_chunks_' (as was
happening previously). The file is only closed when it gets GCd by upper layers
(and deleted) as was happening earlier. This might show up as the process
leaking file descriptors, but is not actually a leak since the files will
eventually be closed on GC

Test Plan:
unit tests on fbcode and integration tests with mtr

Reviewers: arahut

Subscribers:

Tasks:

Tags: