KUDU-3210 Add option to enforce FIPS approved mode
If KUDU_REQUIRE_FIPS_MODE is set to "1", "true" or "yes", Kudu checks if
FIPS mode is enabled and crashes when initializing OpenSSL if not. This
is also true for Kudu C++ client applications.

On initializing OpenSSL, Kudu now also verbose logs (level 2) if FIPS
mode is enabled regardless of whether it's required or not.

Change-Id: I98a6a8b3330ea0b372b188690fadd4d312d8bf93
Reviewed-on: http://gerrit.cloudera.org:8080/16657
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins