[common] introduce TxnId wrapper
TxnId is a wrapper around int64_t with transaction-related
specifics.  It's an alternative to boost::optional<int64_t> since
it lightweight and it's possible to use it as an atomic
(i.e. std::atomic<TxnId> works as expected).

This patch also contains appropriate test coverage for the newly
introduced functionality.

Change-Id: I0511637702e946f140dc617ee6bf7e1c3d5289e3
Reviewed-on: http://gerrit.cloudera.org:8080/16624
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>