[tools] mention whitespace as a separator for variadic args
I witnessed a case when people try to use a comma as separators for
a variadic argument in a kudu CLI command.  As it turned out, the
help/usage message doesn't state clearly that multiple items in a
variadic argument should be separated by a whitespace.  This changelist
updates the description of existing variadic arguments to explicitly
mention that the whitespace is the expected separator there.

This patch does not contain any functional changes.

Change-Id: I3ef979bd17655b5d46e1adb2a2f22923353ac435
Reviewed-on: http://gerrit.cloudera.org:8080/16642
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins