[kuduraft] Do not return early from LogCache::ReadOps()
Summary:
LogCache::ReadOps() returns early if the Op to be read is not found in the index
(after calling LookupOpId()). This is wrong for multiple reasons:
1. There is custom code added in the upper layer (plugin) to trap errors like
NotFound() and report it correctly to the error manager so that appropriate
action can be taken (either evict the peer that is trying to catchup or change
the leader to the node that might still have the logs required by a particular
peer to catchup)
2. In the very rear case, index might not have the OpId, but LogCache could
still have it cached

This PR tries to handle this case. If LookupOpId() returns a NotFound() error,
we make a dummy call into log_->ReadReplicatesInRange() to read a single Op so
that it gets a chance to update error manager. We then return immediately from
LogCache::ReadOps() to preserve existing behavior.

Test Plan: tests added in fbcode to verify that appropriate error is reported

Reviewers: arahut

Subscribers:

Tasks:

Tags: