KUDU-1563. Add UPDATE_IGNORE and DELETE_IGNORE operations
This patch adds an UPDATE_IGNORE operation which behaves like
a normal UPDATE except a key not found error will not be raised by the
primary key having not been previously inserted.

This patch also adds a DELETE_IGNORE operation which behaves like
a normal DELETE except a key not found error will not be raised by the
primary key having not been previously inserted.

Follow on patches will add more client support.

Note: This patch is a follow on to https://gerrit.cloudera.org/#/c/4491/.

Change-Id: I11dfd06e8d4d22cf1097fe1ff01a1b97cafaf899
Reviewed-on: http://gerrit.cloudera.org:8080/16661
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>