KUDU-2612 C++ client transaction API
This patch adds C++ client API to interact with TxnManager.
Corresponding tests are added as well.  True end-to-end tests are not
available yet because they require transaction orchestration
functionality to be present, but it's not there yet.  I'm planning
to add more tests as soon as the orchestration starts to appear.

This patch is mostly focused on the client-side transactional API,
containing several TODOs.  The scope of those is limited to the
related client-side-only functionality.  I'm planning to address
those in follow-up patches.  I think that way it's easier to scope,
implemented and review compared with lumping them all in a single patch.

Change-Id: Ic48233f72873012ea36ff4a05d16c58a0ba9b407
Reviewed-on: http://gerrit.cloudera.org:8080/16710
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>