[client] KUDU-2966 configure connection negotiation timeout
This patch addresses KUDU-2966 for the Kudu C++ client, making
the timeout for the connection negotiation configurable.  This patch
also contains a unit test for the newly introduced functionality.

As for the motivation for this patch: we have observed a few clusters
in the wild where connection negotiations were a bit slow when servers
are under heavy load.  It would be nice to control the connection
negotiation timeout at the client side as well, so at least kudu CLI
tool could use this functionality.

The newly introduced functionality is to be used at least in kudu CLI
(will be posted in a separate follow-up patch).

Change-Id: I8573a572887be041637e28e518a7cd3a6d1f1693
Reviewed-on: http://gerrit.cloudera.org:8080/16687
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>