[consensus][proxy] fix default proxy routing
This patch makes the behavior of a proxy node more reasonable when the
leader has a proxy config defined but the proxy node does not.
Previously, that type of mismatch would result in a routing loop. Now,
it results in the proxy node falling back to direct routing.

Includes a unit test.