[server] fix compilation under RHEL/CentOS 6
It seems the version of the OpenSSL library used to work around
versioned symbols in RHEL/CentOS 6.5+ doesn't contain declaration
of FIPS_mode() function in openssl/crypto.h file.  That breaks the
compilation after the introduction of the two changelist mentioned
below.  This patch fixes the compilation breakage, adding
openssl/fips.h header file when necessary.

This is a follow-up to changelists 46231c52f and 6d2c79a4f.

Change-Id: I3714365801f7de4e4768f4c4e2cf063ff5a65c57
Reviewed-on: http://gerrit.cloudera.org:8080/16696
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins