[tablet_copy_client-test] fix scenario on DownloadBlocks() error reporting
In patch [1] the status reporting of the multi-threaded DownloadBlocks()
was not tested properly. The original test scenario called only
StartCopy(), but DownloadBlocks() was not called.

This patch moves the corresponding test code into a separate scenario,
targeting testing failure paths in any of the download threads spawned
from TabletCopyClient::DownloadBlocks(). With this patch,
both StartCopy() and DownloadBlocks() are called in the scenario.

[1] https://gerrit.cloudera.org/#/c/16274/8/src/kudu/tserver/tablet_copy_client-test.cc

Change-Id: I3bfbe12bf13d0cc274ad266f62c5539f7da791b8
Reviewed-on: http://gerrit.cloudera.org:8080/16709
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>